fix: export config at `/config` instead of `/`
Partly aligns with vike-react.

BREAKING CHANGE: Update Vike to `0.4.160` or above.

BREAKING CHANGE: Replace `import vikeVue from 'vike-vue'`
with `import vikeVue from 'vike-vue/config'`. (Typically
in your `/pages/+config.h.js`.)

BREAKING CHANGE: Replace
`import type { OnCreateAppSync } from 'vike-vue'` (or
another available hook type) with
`import type { OnCreateAppSync } from 'vike-vue/types'`
