+++ examples/basic/package.json
@@ -14,7 +14,7 @@
     "vike-vue": "0.6.0",
     "vite": "^5.0.12",
     "unplugin-vue-markdown": "^0.25.2",
-    "vike": "0.4.159",
+    "vike": "^0.4.160",
     "vue": "^3.4.15"
   },
   "type": "module"
