+++ examples/ssr-spa/package.json
@@ -10,7 +10,7 @@
     "typescript": "^5.3.3",
     "vike-vue": "0.6.0",
     "vite": "^5.0.12",
-    "vike": "0.4.159",
+    "vike": "^0.4.160",
     "vue": "^3.4.15"
   },
   "type": "module"
