+++ examples/with-vike-pinia/package.json
@@ -9,7 +9,7 @@
     "@vitejs/plugin-vue": "^5.0.3",
     "pinia": "^2.1.7",
     "typescript": "^5.3.3",
-    "vike": "0.4.159",
+    "vike": "^0.4.160",
     "vike-pinia": "0.0.1",
     "vike-vue": "0.6.0",
     "vite": "^5.0.12",
