+++ examples/with-vue-plugin/pages/+config.h.ts
@@ -1,7 +1,7 @@
 import type { Config } from 'vike/types'
 import Layout from '../layouts/LayoutDefault.vue'
 import Head from '../layouts/HeadDefault.vue'
-import vikeVue from 'vike-vue'
+import vikeVue from 'vike-vue/config'
 
 // Default configs (can be overridden by pages)
 export default {
