+++ packages/vike-pinia/package.json
@@ -35,7 +35,7 @@
     }
   },
   "peerDependencies": {
-    "vike": "^0.4.159",
+    "vike": "^0.4.160",
     "pinia": "^2.0.0",
     "vike-vue": "^0.5.3",
     "vue": "^3.0.0"
@@ -45,7 +45,7 @@
     "@types/node": "^20.11.5",
     "pinia": "^2.1.7",
     "typescript": "^5.3.3",
-    "vike": "0.4.159",
+    "vike": "^0.4.160",
     "vike-vue": "0.6.0",
     "vue": "^3.4.15"
   },
