+++ packages/vike-vue/package.json
@@ -2,14 +2,15 @@
   "name": "vike-vue",
   "version": "0.6.0",
   "type": "module",
-  "main": "./dist/renderer/+config.js",
-  "types": "./dist/renderer/+config.d.ts",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "exports": {
-    ".": "./dist/renderer/+config.js",
+    ".": "./dist/index.js",
+    "./config": "./dist/+config.js",
     "./renderer/onRenderHtml": "./dist/renderer/onRenderHtml.js",
     "./renderer/onRenderClient": "./dist/renderer/onRenderClient.js",
-    "./usePageContext": "./dist/components/usePageContext.js",
-    "./useData": "./dist/components/useData.js",
+    "./usePageContext": "./dist/hooks/usePageContext.js",
+    "./useData": "./dist/hooks/useData.js",
     "./ClientOnly": "./dist/components/ClientOnly.js"
   },
   "dependencies": {},
@@ -21,7 +22,7 @@
     "release:commit": "release-me --git-prefix vike-vue --changelog-dir packages/vike-vue/ commit"
   },
   "peerDependencies": {
-    "vike": "^0.4.159",
+    "vike": "^0.4.160",
     "vue": "^3.0.0"
   },
   "devDependencies": {
@@ -29,7 +30,7 @@
     "@types/node": "^20.11.5",
     "@vitejs/plugin-vue": "^5.0.3",
     "typescript": "^5.3.3",
-    "vike": "0.4.159",
+    "vike": "^0.4.160",
     "vite": "^5.0.12",
     "vite-plugin-dts": "^3.7.1",
     "vue": "^3.4.15"
@@ -37,13 +38,19 @@
   "typesVersions": {
     "*": {
       ".": [
-        "./dist/renderer/+config.d.ts"
+        "./dist/index.d.ts"
+      ],
+      "config": [
+        "./dist/+config.d.ts"
+      ],
+      "types": [
+        "./dist/types/index.d.ts"
       ],
       "usePageContext": [
-        "./dist/components/usePageContext.d.ts"
+        "./dist/hooks/usePageContext.d.ts"
       ],
       "useData": [
-        "./dist/components/useData.d.ts"
+        "./dist/hooks/useData.d.ts"
       ],
       "renderer/onRenderHtml": [
         "./dist/renderer/onRenderHtml.d.ts"
