+++ packages/vike-vue/src/renderer/onRenderHtml.ts
