+++ packages/vike-vue/src/utils/assert.ts
