+++ packages/vike-vue/src/utils/objectAssign.ts
